Add an unexported matrix3x3 type for internal use.
S2 uses some matrix transforms under the covers. In order to ease the
porting of some of the lower level methods in S2, define a simple 3x3
matrix type. This type is bare bones and contains only methods used by
s2. It is not a complete implementation of a matrix. The main
differences are that a general matrix type would take and return slices
or arrays, whereas this takes and returns s2.Points.

Signed-off-by: David Symonds <dsymonds@golang.org>