s2: Improve implementation of Cell.ContainsPoint.
Changed the implementation of Cell.ContainsPoint() to ensure that when a
point is converted to a CellId, the returned S2CellId always contains
that point. In other words,

  CellFromCellID(cellIdFromPoint(p)).ContainsPoint(p)

is always true. Previously, there were rare cases where this could
fail for points near the edges of a Cell.

Also updated the documentation to clarify that S2Cells are considered
to be closed sets. In other words, points along the edge of a Cell
are considered to belong to the adjacent Cell(s) as well.

Signed-off-by: David Symonds <dsymonds@golang.org>