s2: Fix a bug in Polygon.initLoopProperties.
Before this change to initLoopProperties to start the bounds as an
EmptyRect instead of it defaulting to the zero rect, the bounds
computations failed. This is because C++ gets default constructors on
Rect which default to Empty.

The bounds Union checks were doing Lo checks of the Min of 0 and the
other, and Hi would to Max of 0 and the other. So polygons that had low
corners above zero or high corners below zero would end up with
oversized bounds.

The same logic on updating bounds was in decodeCompressed. The TODO
there was to do the right way once it was addded which it was a while
back.

Signed-off-by: David Symonds <dsymonds@golang.org>