s2: Update EdgeTesselator to be more robust and conservative.
Previously the algorithm could not handle points of inflection (that is,
edges that cross the equator in the Plate Carree and Mercator
projections), and was not always conservative even when edges did not
have a point of inflection.

The new algorithm solves these problems by modeling the error as a cubic
polynomial. This is done by evaluating the distance between the edges at
two equally space points rather than at the midpoint. This results is a
small amount of extra tessellation (about 1.5% on average) but is much
more robust.

Signed-off-by: David Symonds <dsymonds@golang.org>