edge_query: use shapeID instead of index.idForShape lookup (#136)
Performance of maybeAddResult is not great since it invokes idForShape twice.

idForShape iterates through all the shapes in shapeIndex to find id for
the shape. If the ShapeIndex contains a big number of shapes it can be
quite a big task to find the id for the shape.

Since shapeID is available for maybeAddResult use it instead of idForShape.