bits: Remove special cases and go 1.8 support (#187)
Reimplement findMSBSetNonZero64 with bits.Len64, which is a more direct
reflection of its semantics than bits.LeadingZeros64.

findMSBSetNonZero64 and findLSBSetNonZero64 were called "NonZero"
because their precondition in C++ was that the argument was non-zero.
Change the zero handling of these functions to be more uniform since
it's never used and we don't care.

In a future commit, these functions can be inlined completely.

Fixes https://github.com/golang/geo/issues/179.