rect.CapBound: Fix for tall and wide rects (#233)
For tall (>90 degrees) and wide (>180 degrees) lat/lng rects,
rect.CapBound returned an incorrect cap that didn't contain the full
rect. The function intends to return a polar cap in this case, but due
to an apparent typo (2 * Pi instead of Pi), returned a cap centered at
the center of the rect instead.

This fix has recently been applied to C++ and Java in google3, but has
not been exported to those github repos yet.